refactor(harness): extract doctor drift helpers under size ratchet (#1218)
Move expected-state comparison and receipt attestation out of
harness_profile_cmd.py so the allowlisted module stays at or below
2570 lines without raising the ceiling.
